backend/requirements.txt
aiohttp
python-socketio
openai
edge-tts
cryptography
colorama
